lookupserv/requirements.txt
PyYAML==3.11
pyzmq==14.7.0
wsgiref==0.1.2
